+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/entity/Podcast.kt
@@ -103,6 +103,7 @@ data class Podcast(
     @ColumnInfo(name = "slug") var slug: String = "",
     @ColumnInfo(name = "explicit") var explicit: Boolean? = null,
     @ColumnInfo(name = "web_feed", defaultValue = "0") var webFeed: Boolean = false,
+    @ColumnInfo(name = "network_list_id") var networkListId: String? = null,
     @Embedded(prefix = "bundle") var singleBundle: Bundle? = null,
     @Ignore val episodes: MutableList<PodcastEpisode> = mutableListOf(),
 ) : Serializable {