+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/to/ImprovedSearchResultItem.kt
@@ -25,6 +25,13 @@ sealed interface ImprovedSearchResultItem {
         val isExplicit: Boolean = false,
     ) : ImprovedSearchResultItem
 
+    data class NetworkItem(
+        override val uuid: String,
+        override val title: String,
+        val description: String? = null,
+        val imageUrl: String? = null,
+    ) : ImprovedSearchResultItem
+
     data class EpisodeItem(
         override val uuid: String,
         override val title: String,